+++ src/plugin/managers/shape_manager/core.py
@@ -2,39 +2,37 @@ import logging
 from collections.abc import Mapping
 from multiprocessing import Pool
 
+from plugin.config import (
+    RETRIEVE_SHAPE_CONFIG,
+)
 from plugin.presentation.callbacks import AbstractProgressCallback
-from spectrumlab.emulations.noise import Noise
-from spectrumlab.peaks.shape import (
-    DraftPeakConfig,
-    RestoreShapeConfig,
+from spectrumlab.peaks import (
+    draft_peaks,
+)
+from spectrumlab.shapes import (
     Shape,
-    restore_shape_from_spectrum,
-    FIGURES,
-    FIGURE,
+    retrieve_shape_from_spectrum,
 )
 from spectrumlab.spectra import Spectrum
 
 LOGGER = logging.getLogger('plugin-peak-shape')
 
 
-def restore_shape(__args) -> Shape:
-    n, spectrum, draft_peak_config, restore_shape_config = __args
+def retrieve_shape(__args) -> Shape:
+    n, spectrum  = __args
 
     LOGGER.debug(
-        'detector %02d - restore peak\'s shape',
-        n,
+        'detector %02d - retrieve peak\'s shape',
+        n+1,
     )
     try:
-        FIGURE.set(FIGURES.get()[n])
-
-        shape = restore_shape_from_spectrum(
+        peaks = draft_peaks(
             spectrum=spectrum,
-            noise=Noise(
-                detector=spectrum.detector,
-                n_frames=15000,  # TODO: read from xml!
-            ),
-            draft_peak_config=DraftPeakConfig(**draft_peak_config.model_dump()),
-            restore_shape_config=RestoreShapeConfig(**restore_shape_config.model_dump()),
+        )
+        shape = retrieve_shape_from_spectrum(
+            spectrum=spectrum,
+            peaks=peaks,
+            n=n,
         )
 
         # FIXME: Atom's lagacy (never change it!)
@@ -44,80 +42,61 @@ def restore_shape(__args) -> Shape:
             ratio=(1 - shape.ratio),
         )
 
-    except ValueError as error:
+    except Exception as error:
         LOGGER.warning(
-            'detector %02d - shape is not restored: %r',
-            n,
+            'detector %02d - shape is not retrieved: %r',
+            n+1,
             error,
         )
-        LOGGER.debug(
-            'detector %02d - default shape is used: %r',
-            n,
-            restore_shape_config.default_shape,
-        )
-        return restore_shape_config.default_shape
+        return RETRIEVE_SHAPE_CONFIG.default_shape
 
-    LOGGER.debug(
-        'Restored peak\'s shape for detector %d: %s',
-        n,
+    LOGGER.info(
+        'detector %02d - shape is retrieved: %s',
+        n+1,
         shape,
     )
     return shape
 
 
-def restore_shapes(
+def retrieve_shapes(
     n_workers: int,
     spectra: Mapping[int, Spectrum],
     progress_callback: AbstractProgressCallback,
-    draft_peak_config,
-    restore_shape_config,
-) -> tuple[Shape]:
+) -> Mapping[int, Shape]:
 
     if n_workers > 1:
-        return restore_shapes_multiprocess(
+        return retrieve_shapes_multiprocess(
             n_workers=n_workers,
             spectra=spectra,
             progress_callback=progress_callback,
-            draft_peak_config=draft_peak_config,
-            restore_shape_config=restore_shape_config,
         )
 
     shapes = {}
     for n, spectrum in spectra.items():
-        shape = restore_shape((n, spectrum, draft_peak_config, restore_shape_config))
+        shape = retrieve_shape((n, spectrum))
 
-        progress_callback(
-            n=n,
-            total=len(spectra),
-        )
         shapes[n] = shape
+        progress_callback(n=n)
 
     return shapes
 
 
-def restore_shapes_multiprocess(
+def retrieve_shapes_multiprocess(
     n_workers: int,
     spectra: Mapping[int, Spectrum],
     progress_callback: AbstractProgressCallback,
-    draft_peak_config,
-    restore_shape_config,
-) -> tuple[Shape]:
+) -> Mapping[int, Shape]:
 
-    shapes = []
+    shapes = {}
     with Pool(n_workers) as pool:
-        for shape in pool.imap(
-            restore_shape,
+        for n, shape in enumerate(pool.imap(
+            retrieve_shape,
             [
-                (n, spectrum, draft_peak_config, restore_shape_config)
-                for n, spectrum in enumerate(spectra)
+                (n, spectrum)
+                for n, spectrum in spectra.items()
             ],
-        ):
-            shapes.append(shape)
+        )):
+            shapes[n] = shape
+            progress_callback(n=n)
 
-            progress_callback(
-                n=len(shapes),
-                total=len(spectra),
-            )
-            shapes.append(shape)
-
-    return tuple(shapes)
+    return shapes