+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 * logging for spectra parsers
-* progress bar
+* progress bar window
 
 
 ## [0.1.2] - 2025-02-25
@@ -28,5 +28,22 @@ All notable changes to this project will be documented in this file.
 ## [0.2.0] - 2025-03-23
 
 ### Added
-* observability window
+* view window
 * config for peak's drafting and peak's shape restoring
+
+
+## [0.3.0] - 2025-03-30
+
+### Refac
+* update plugin structure
+
+### Added
+* peak filtration by width
+* n_frames and clipped
+* filtration in drafting (supposed by Stas Dodonov)
+
+### Fixed
+* progress bar window
+* multiprocessing with progress bar window
+* default shape
+* overlay spectrum and shape figures